requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
pydantic-settings==2.1.0
slowapi==0.1.9
structlog==24.1.0
python-multipart==0.0.22
httpx==0.26.0
